+++ tests/event-bus-group.test.ts
@@ -1,4 +1,4 @@
-import { EventBusGroup, EventBusRxJs as EventBus } from "@mod";
+import { EventBusGroup, EventBusRxJS as EventBus } from "@mod";
 import { EventWithPayload, PlainEvent } from "@testdata/events.mock.ts";
 import { assertEquals, assertExists, describe, it } from "@deps/std_testing.ts";
 import { assertError } from "@testutils/assertError.ts";